requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
defusedxml==0.7.1
EbookLib==0.18
et_xmlfile==2.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
joblib==1.4.2
lml==0.1.0
lxml==5.3.0
numpy==2.2.2
odfpy==1.4.1
ollama==0.4.7
openpyxl==3.1.5
pandas==2.2.3
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
pyexcel-io==0.6.7
pyexcel-ods==0.6.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-pptx==1.0.2
pytz==2024.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
striprtf==0.0.28
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
XlsxWriter==3.2.2
